Fixed color code revering issues and allow it to run on python 3.8.x
Fixed(?) how smi style color code converted to ass style.
Changed and test to run on python 3.8.x on linux_x86-64 and arm Also, changed to not add "kor" on the end of the file when there is no other language is depicted.
Also, add to print what file is currently working on.
Also, some setting update.